+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed
+
+- Link base instance options to server and map options
+
 ## [0.5.3] 2025-10-29
 ## [0.5.2] 2025-10-29
 ## [0.5.1] 2025-10-04